fix(kv): derive KV byte totals from real allocations — count the GPU ring (#246) (#258)
* fix(kv): derive KV byte totals from real allocations — count the GPU ring (#246)

`kv_bytes` could not see the GPU ring that the ring-backed K codecs stand up
at decode. It was worse than a missing buffer: two accountings existed for the
same memory and had drifted apart.

- `KvStorage::resident_bytes` routed through a set of per-codec
  `quant_*_bytes` free functions that counted CPU blocks only.
- Each store also had its own `byte_size`, which did count the ring — but
  nothing in production called it. It was reachable only from tests, so the
  two mirrors were free to disagree, and did.
- `KvCache::approx_bytes` — the source of the `kv_bytes` trace event — was a
  per-codec bits-per-element formula that never read the store at all. On
  Bonsai-8B it reported byte-identical KV for `k_iso3`, `k_rotor3` and `k8v4`:
  three codecs, three different storage layouts, one number.

A nominal bit-width is not a cache's memory. Only the store knows what it
allocated, so the store is now the single source of the figure:

  KvCache::resident_bytes -> KvStorage::resident_bytes -> <Store>::byte_size

`approx_bytes` and the `quant_*_bytes` mirror are both deleted (-366 lines of
formula). Every `byte_size` is built from two primitives in the new
`rmlx-kv-quant::bytes` module — `Array` shape x dtype, and `Vec` len x elem
size — so a buffer that grows, or changes dtype, reports the truth with
nothing to update. This also picks up buffers the old mirror never counted at
all: `QuantV`'s TCQ codebook and high-precision indices, and the ring behind
`QuantIsoK3/K4` and `QuantRotorK3/K4`. `LinearAttnCache` loses its hard-coded
2-byte/4-byte item sizes for the same reason.

Anti-drift is compiler-enforced, not documented: every `byte_size` and
`resident_bytes` opens with an exhaustive `let Self { .. }` destructure naming
all fields, and `KvStorage::resident_bytes` binds every variant field with no
`..` rest-patterns. Adding a buffer anywhere in the chain is E0027 until it is
classified as payload or metadata. Mutation-checked at three levels
(`QuantIsoK3`, `QuantV`, `KvCache`): each probe went red, then was reverted.

Measured (Bonsai-8B `Qwen3ForCausalLM`, decode-time trace, bytes/layer):

  ctx  codec     rings   before        after       delta
  4k   k_iso3      26    25,288,704    42,902,144  +69.6%
  4k   k_rotor3     0    25,288,704    29,042,864  +14.8%
  4k   k8v4         0    25,288,704    32,923,648  +30.2%
  16k  k_iso3      26   100,786,176   169,517,696  +68.2%

Anchored to reality rather than to the new formula: on Bonsai-8B `k_iso3` @4k
the reported KV total accounted for 60% of the process's RSS growth before
(0.85 GiB claimed vs 1.42 GiB grown) and 99% after (1.44 GiB vs 1.45 GiB).

Gemma-4-e2b isolates the ring on its own: its metric already used
`resident_bytes` on both sides, so the only delta is the ring — `k_iso3`
+23.6% @4k / +24.7% @16k (3 ring-backed layers), while `k_rotor3` and `k8v4`
move by exactly 0.0%, confirming no perturbation where there is no ring.

Tests: `resident_ring_tests` asserts the reported total grows by at least the
ring's own measured size across both ring-backed codecs and two contexts —
against the ring's real allocation, never a restatement of the accounting.
The warm-TTFT and models-side byte tests are re-anchored to the live buffers
instead of the deleted formula's literal output.

Docs: KV_QUANT / KV_CACHE / PERF_BASELINE corrected. The Bonsai-27B KV-MB
column is marked SUSPECT for the four ring-backed cells (`k_iso3`, `k_iso4`,
`k_rotor3`, `k_rotor4`) rather than left standing as fact — re-measurement is
a separate job.

* fix(kv): emit kv_bytes once per request; close the RotatingState and Paged guard gaps

Review follow-ups.

**`--log verbose` was made quadratic.** The nine per-layer `kv_bytes` trace
sites swapped an O(1) formula for `resident_bytes`, which walks a block list
that gains an entry per decode step — so enabling the diagnostic made decode
O(n_layers x N^2/2) in context. Level gating hides an event's emission cost,
not the cost of computing its fields.

The nine sites are deleted. `kv_bytes` is now emitted once per request from
`PromptCache::store_kv_cache_bytes`, which every arch already calls with the
sum pre-computed — so the event costs nothing at all, and carries the
whole-model total instead of an unattributable per-layer figure (the old event
had no `layer_idx`). Event count per request: 864 -> 1. The O(blocks) cost is
now stated on `resident_bytes` itself, with a note not to "fix" it with a
cached counter — that would rebuild the mirror this work deleted.

**Guard gap at `RotatingState`** (the Gemma4 SWA path). `KvCache::resident_bytes`
destructured its own fields but then reached the ring's buffers by field access,
so a third buffer added to `RotatingState` would have compiled clean and gone
uncounted — the exact class the guard exists to stop. It now has its own
destructured `byte_size(filled)`. Mutation-checked: E0027 at `rotating.rs:80`,
reverted.

**Surviving mirror in the Paged subtree.** `PageSlab::resident_bytes` restated
geometry (`pages x page_tokens x elems_per_token x itemsize`) instead of reading
each page's `Array`; `pool` was never consulted for its shape. Now
`pool.iter().flatten().map(array_bytes).sum()`. The three `Paged*Storage`
bodies gain the destructure guard. Numerically identical today (`--paged-kv` is
default-OFF) — this closes the drift, and makes the "no per-codec byte formula"
claim true rather than aspirational.

**Third copy of the primitive.** `resident_bytes` opened with a local `ab()`
byte-identical to `crate::bytes::array_bytes`. Deleted; `filled_seq_bytes`
moved into `crate::bytes` where its second caller (`RotatingState`) needs it.

**Tests claimed more than they delivered.** The ring tests anchor on
`QuantKGpuRing::byte_size`, which *is* the accounting under test — so they prove
the ring is plumbed into the total (the shipped bug), not that its magnitude is
right. Docstrings now say that. Added `ring_bytes_match_independent_geometry`,
which rebuilds the ring's size from its documented layout (codes u32
[cap*kv_h*n_groups] + scales f32 [same] + norms f32 [cap*kv_h]) and compares.
Mutation-checked: dropping the norms buffer fails it (1,048,576 vs 1,064,960 B);
the delta tests sail through that.

**Docs.** `KV_QUANT.md` no longer states as present fact that `resident_bytes`
is what `kv_cache_bytes` records everywhere — it now tables the per-arch sample
point (gemma4 / qwen3.5-moe / qwen3-exact-hit sample after decode and see the
ring; qwen3-miss / qwen2 / laguna sample pre-decode and cannot), and notes
unification is tracked separately. The paged-pool-overhead caveat, orphaned when
`approx_bytes` was deleted, is re-homed onto `KvStorage::resident_bytes`. The
27B note says "unaffected within reading precision" and names the TCQ-sideband
exception. `CLAUDE.md`'s trace-level example is corrected: `kv_bytes` is no
longer a per-layer `trace!`.

Behaviour-neutral: gemma-4-e2b `k_iso3` @4k still reports 70,290,624 B and
`k8v4` 42,627,072 B; Bonsai-8B `k_iso3` @4k still 1,138,821,696 B — unchanged
across the hoist and the RotatingState delegation. Gemma is the SWA-heavy arch,
so it exercises the new ring path.

+++ crates/rmlx-kv-quant/src/storage/quant_rotor_v4_tests.rs
@@ -82,12 +82,12 @@ fn quant_rotor_v4_byte_size_counts_rotors_once() {
     let data = vec![0.5_f32; 1 * 1 * 4 * head_dim];
     qv.append(&data, &new_shape).unwrap();
 
-    let rotors_bytes = qv.rotors.len() * size_of::<f32>();
+    let rotors_bytes = (qv.rotors.len() * size_of::<f32>()) as u64;
     let blocks_bytes = qv
         .blocks
         .iter()
-        .map(|b| b.codes.len() * 4 + b.scales.len() * 4 + b.norms.len() * 4)
-        .sum::<usize>();
+        .map(|b| (b.codes.len() * 4 + b.scales.len() * 4 + b.norms.len() * 4) as u64)
+        .sum::<u64>();
     assert_eq!(qv.byte_size(), rotors_bytes + blocks_bytes);
 
     // Appending a second block must not double-count rotors.
@@ -95,8 +95,8 @@ fn quant_rotor_v4_byte_size_counts_rotors_once() {
     let new_blocks_bytes = qv
         .blocks
         .iter()
-        .map(|b| b.codes.len() * 4 + b.scales.len() * 4 + b.norms.len() * 4)
-        .sum::<usize>();
+        .map(|b| (b.codes.len() * 4 + b.scales.len() * 4 + b.norms.len() * 4) as u64)
+        .sum::<u64>();
     assert_eq!(qv.byte_size(), rotors_bytes + new_blocks_bytes);
 }
